+++ starter/starter-adapter/starter-adapter-webserver/src/main/java/org/dromara/dynamictp/starter/adapter/webserver/adapter/undertow/UndertowDtpAdapter.java
@@ -15,9 +15,9 @@
  * limitations under the License.
  */
 
-package org.dromara.dynamictp.starter.adapter.webserver.autocconfigure.spring.undertow;
+package org.dromara.dynamictp.starter.adapter.webserver.adapter.undertow;
 
-import org.dromara.dynamictp.starter.adapter.webserver.autocconfigure.spring.AbstractWebServerDtpAdapter;
+import org.dromara.dynamictp.starter.adapter.webserver.adapter.AbstractWebServerDtpAdapter;
 import org.dromara.dynamictp.common.properties.DtpProperties;
 import org.dromara.dynamictp.common.util.ReflectionUtil;
 import org.dromara.dynamictp.core.support.ExecutorWrapper;
