+++ starter/starter-adapter/starter-adapter-webserver/src/main/java/org/dromara/dynamictp/starter/adapter/webserver/adapter/undertow/taskpool/ExecutorServiceTaskPoolAdapter.java
@@ -15,27 +15,25 @@
  * limitations under the License.
  */
 
-package org.dromara.dynamictp.starter.adapter.webserver.autocconfigure.spring.undertow.taskpool;
+package org.dromara.dynamictp.starter.adapter.webserver.adapter.undertow.taskpool;
 
-import org.dromara.dynamictp.starter.adapter.webserver.autocconfigure.spring.undertow.UndertowTaskPoolEnum;
+import org.dromara.dynamictp.starter.adapter.webserver.adapter.undertow.UndertowTaskPoolEnum;
 import org.dromara.dynamictp.core.support.ExecutorAdapter;
 import org.dromara.dynamictp.core.support.ThreadPoolExecutorAdapter;
 
 import java.util.concurrent.ThreadPoolExecutor;
 
-import static org.dromara.dynamictp.starter.adapter.webserver.autocconfigure.spring.undertow.UndertowTaskPoolEnum.THREAD_POOL_EXECUTOR_TASK_POOL;
-
 /**
- * ThreadPoolExecutorTaskPoolHandler related
+ * ExecutorServiceTaskPoolAdapter related
  *
  * @author yanhom
  * @since 1.1.3
  */
-public class ThreadPoolExecutorTaskPoolAdapter implements TaskPoolAdapter {
+public class ExecutorServiceTaskPoolAdapter implements TaskPoolAdapter {
 
     @Override
     public UndertowTaskPoolEnum taskPoolType() {
-        return THREAD_POOL_EXECUTOR_TASK_POOL;
+        return UndertowTaskPoolEnum.EXECUTOR_SERVICE_TASK_POOL;
     }
 
     @Override
