+++ starter/starter-adapter/starter-adapter-webserver/src/main/java/org/dromara/dynamictp/starter/adapter/webserver/adapter/undertow/taskpool/TaskPoolAdapter.java
@@ -15,31 +15,33 @@
  * limitations under the License.
  */
 
-package org.dromara.dynamictp.starter.adapter.webserver.autocconfigure.spring.undertow.taskpool;
+package org.dromara.dynamictp.starter.adapter.webserver.adapter.undertow.taskpool;
 
-import org.dromara.dynamictp.starter.adapter.webserver.autocconfigure.spring.undertow.UndertowTaskPoolEnum;
+import org.dromara.dynamictp.starter.adapter.webserver.adapter.undertow.UndertowTaskPoolEnum;
 import org.dromara.dynamictp.core.support.ExecutorAdapter;
-import org.dromara.dynamictp.core.support.ThreadPoolExecutorAdapter;
 
-import java.util.concurrent.ThreadPoolExecutor;
-
-import static org.dromara.dynamictp.starter.adapter.webserver.autocconfigure.spring.undertow.UndertowTaskPoolEnum.EXECUTOR_SERVICE_TASK_POOL;
+import java.util.concurrent.Executor;
 
 /**
- * ExecutorServiceTaskPoolAdapter related
+ * TaskPoolAdapter related
  *
  * @author yanhom
  * @since 1.1.3
  */
-public class ExecutorServiceTaskPoolAdapter implements TaskPoolAdapter {
+public interface TaskPoolAdapter {
 
-    @Override
-    public UndertowTaskPoolEnum taskPoolType() {
-        return EXECUTOR_SERVICE_TASK_POOL;
-    }
+    /**
+     * Get the task pool type
+     *
+     * @return task pool type
+     */
+    UndertowTaskPoolEnum taskPoolType();
 
-    @Override
-    public ExecutorAdapter<ThreadPoolExecutor> adapt(Object executor) {
-        return new ThreadPoolExecutorAdapter((ThreadPoolExecutor) executor);
-    }
+    /**
+     * Adapt the task pool
+     *
+     * @param taskPool task pool
+     * @return executor adapter
+     */
+    ExecutorAdapter<? extends Executor> adapt(Object taskPool);
 }
